all collectors finished with authentication

+++ kuhako/src/main/java/com/capstone/kuhako/services/CollectorModuleServices/SendFollowUpServiceImpl.java
@@ -1,8 +1,10 @@
 package com.capstone.kuhako.services.CollectorModuleServices;
 
 
+import com.capstone.kuhako.models.Collector;
 import com.capstone.kuhako.models.CollectorModules.SendFollowUp;
 import com.capstone.kuhako.repositories.CollectorModuleRepository.SendFollowUpRepository;
+import com.capstone.kuhako.repositories.CollectorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -12,30 +14,50 @@ import org.springframework.stereotype.Service;
 public class SendFollowUpServiceImpl implements SendFollowUpService{
     @Autowired
     private SendFollowUpRepository sendFollowUpRepository;
+    @Autowired
+    private CollectorRepository collectorRepository;
 
-    public void createSendFollowUp(SendFollowUp sendFollowUp){
-//      Collector author = collectorRepository.findByCollectorname(jwtToken.getCollectornameFromToken(stringToken));
-        SendFollowUp newSendFollowUp = new SendFollowUp();
-        newSendFollowUp.setCollector(sendFollowUp.getCollector());
-        newSendFollowUp.setPaymentStatus(sendFollowUp.getPaymentStatus());
-        newSendFollowUp.setReminderMessage(sendFollowUp.getReminderMessage());
-        sendFollowUpRepository.save(newSendFollowUp);
+
+    public void createSendFollowUp(Long collectorId, SendFollowUp sendFollowUp){
+        Collector collector = collectorRepository.findById(collectorId).orElse(null);
+        if (collector != null) {
+            sendFollowUp.setCollector(collector);
+            sendFollowUpRepository.save(sendFollowUp);
+        }
     }
+
     public Iterable<SendFollowUp> getSendFollowUp(){
         return sendFollowUpRepository.findAll();
     }
 
-    public ResponseEntity deleteSendFollowUp(Long id){
-        sendFollowUpRepository.deleteById(id);
-        return new ResponseEntity<>("SendFollowUp Deleted successfully", HttpStatus.OK);
+    public Iterable<SendFollowUp> getSendFollowUpByCollectorId(Long collectorId){
+        return sendFollowUpRepository.findSendFollowUpByCollectorId(collectorId);
     }
 
-    public ResponseEntity updateSendFollowUp(Long id, SendFollowUp sendFollowUp){
-        SendFollowUp sendFollowUpForUpdate = sendFollowUpRepository.findById(id).get();
-        sendFollowUpForUpdate.setCollector(sendFollowUp.getCollector());
-        sendFollowUpForUpdate.setPaymentStatus(sendFollowUp.getPaymentStatus());
-        sendFollowUpForUpdate.setReminderMessage(sendFollowUp.getReminderMessage());
-        sendFollowUpRepository.save(sendFollowUpForUpdate);
-        return new ResponseEntity("SendFollowUp updated successfully", HttpStatus.OK);
+    public ResponseEntity deleteSendFollowUp(Long collectorId, Long id){
+        SendFollowUp sendFollowUpToDelete = sendFollowUpRepository.findById(id).orElse(null);
+
+        if (sendFollowUpToDelete != null && sendFollowUpToDelete.getCollector().getCollector_id().equals(collectorId)) {
+            sendFollowUpRepository.deleteById(id);
+            return new ResponseEntity<>("SendFollowUp Deleted successfully", HttpStatus.OK);
+        } else if (sendFollowUpToDelete != null) {
+            return new ResponseEntity<>("SendFollowUp does not belong to the specified collector", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>("SendFollowUp not found", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    public ResponseEntity updateSendFollowUp(Long collectorId, Long id, SendFollowUp sendFollowUp){
+        SendFollowUp sendFollowUpForUpdate = sendFollowUpRepository.findById(id).orElse(null);
+        if (sendFollowUpForUpdate != null && sendFollowUpForUpdate.getCollector().getCollector_id().equals(collectorId)) {
+            sendFollowUpForUpdate.setPaymentStatus(sendFollowUp.getPaymentStatus());
+            sendFollowUpForUpdate.setReminderMessage(sendFollowUp.getReminderMessage());
+            sendFollowUpRepository.save(sendFollowUpForUpdate);
+            return new ResponseEntity<>("SendFollowUp Updated successfully", HttpStatus.OK);
+        } else if (sendFollowUpForUpdate != null) {
+            return new ResponseEntity<>("SendFollowUp does not belong to the specified collector", HttpStatus.NOT_FOUND);
+        } else {
+            return new ResponseEntity<>("SendFollowUp not found",HttpStatus.NOT_FOUND);
+        }
     }
 }
